backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
langchain-groq
langgraph
pydantic
python-dotenv
faiss-cpu
arxiv
wikipedia
